April 1st 2014 Deadline:
This week we deal a bit of data handling, by continuing from where we left off previously. We set up a query for our users
on Parse.com as well as handled how to execute a query in our code. The method used by parse.com is a bit different from
that of SQLite. After that we created a way in which relationships between friends  can be stored on the the backend. We
then added check marks to  people who are already on our friend's list, to finalize this week's project, we displayed all our
friends under the friends tab.